% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{BEA_DSlist}
\alias{BEA_DSlist}
\title{Get a List of Datasets Available from the BEA API}
\usage{
BEA_DSlist(key)
}
\arguments{
\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov)}
}
\value{
data frame with columns DatasetName and DatasetDescription
}
\description{
\code{BEA_DSlist} returns a data frame with a list of available datasets
}
\details{
Queries the BEA API to get a list of available datasets
}
\examples{
bea.ds <- BEA_DSlist() # don't need to give it a key if you have yours set
}
\keyword{BEA_DSlist}

