% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{BEA_DSparams}
\alias{BEA_DSparams}
\title{Get Parameters for a Particular BEA Dataset}
\usage{
BEA_DSparams(dsname, key)
}
\arguments{
\item{dsname}{text name of the dataset (see \code{BEA_DSlist})}

\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov)}
}
\value{
data frame with columns:
  "ParameterName"
  "ParameterDataType"
  "ParameterDescription"
  "ParameterIsRequiredFlag"
  "MultipleAcceptedFlag"
  "AllValue"
  "ParameterDefaultValue"
}
\description{
\code{BEA_DSparams} returns a data frame with a list of parameters for the particular dataset
Pay particular attention to whether a data set will accept multiple table flags, or whether tables must be
retrieved one by one
}
\details{
Queries the BEA API to get list of parameters needed to retrieve data for a particular dataset. Particularly useful datasets are NIPA and NIUnderlyingDetail
}
\examples{
bea.param <- BEA_DSparams("NIPA") # don't need to give it a key if you have yours set
BEA_DSparams("NIUnderlyingDetail")
BEA_DSparams("RegionalData")
}
\keyword{BEA_DSparams}

