% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{BEA_ParamVals}
\alias{BEA_ParamVals}
\title{Get Parameter Values for a Parameter in a Chosen BEA Dataset}
\usage{
BEA_ParamVals(dsname, pname, key)
}
\arguments{
\item{dsname}{text name of the dataset (e.g., "NIPA") (see \code{BEA_DSlist})}

\item{pname}{text name of the parameter of interest (e.g., "TableID" for the dataset "NIPA")  (see \code{BEA_DSparams})}

\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov)}
}
\value{
data frame with information about allowable values for the parameter of interest
}
\description{
\code{BEA_ParamVals} returns a data frame with a list of parameters for the particular dataset
Pay particular attention to whether a data set will accept multiple table flags, or whether tables must be
retrieved one by one
}
\details{
Queries the BEA API to get allowable values for a parameter of interest for a dataset of interest.
Particularly useful datasets are NIPA and NIUnderlyingDetail
}
\examples{
nipa.tablist <- BEA_ParamVals("NIPA", "TableID") # don't need to give a key if you have yours set
head(nipa.tablist)
# don't need to give a key if you have yours set
nipa.uld.tablist <- BEA_ParamVals("NIUnderlyingDetail", "TableID")
head(nipa.uld.tablist)
BEA_ParamVals("RegionalData", "KeyCode") # TPI_SI is Total personal income (state annual income)
}
\keyword{BEA_ParamVals}

