% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{BEA_RgnData}
\alias{BEA_RgnData}
\title{Get BEA regional data}
\usage{
BEA_RgnData(keycode, key)
}
\arguments{
\item{keycode}{a variable name (see \code{BEA_ParamVals}); default: no default}

\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov); default: my key}
}
\value{
data frame with the regional data
}
\description{
\code{BEA_RgnData} returns a data frame with requested regional data. Right now only works with annual.
}
\details{
Queries the BEA API to get the data. It determines from the keycode whether it is annual or quarterly data
}
\examples{
BEA_DSparams("RegionalData")
geofips <- BEA_ParamVals("RegionalData", "GeoFips")
head(geofips)
BEA_ParamVals("RegionalData", "KeyCode")
keycode <- "GDP_SP"
df <- BEA_RgnData(keycode)
head(df)
# Now get quarterly data
df <- BEA_RgnData("PROP_QI")
head(df)
}
\keyword{BEA_RgnData}

