% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census_api.r
\name{Census_eits}
\alias{Census_eits}
\title{Get a Dataset from the Census eits API}
\usage{
Census_eits(dsname, key)
}
\arguments{
\item{dsname}{text in c("qtax", "qpr", "vip") - quarterly tax, quarterly public retirement systems, value in place}

\item{key}{Your Census API key (can be obtained for free - check www.census.gov)}
}
\value{
data frame
}
\description{
\code{Census_eits} returns a data frame with a list of available datasets
}
\details{
Queries the Census API to get the requested data
}
\examples{
\dontrun{
# don't need to give it a key if you have yours set
cen.qtax <- Census_eits("qtax")
cen.qpr <- Census_eits("qpr")
cen.vip <- Census_eits("vip")
}
}
\keyword{Census_eits}

