% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FRED.r
\name{FRED2}
\alias{FRED2}
\title{Get a series from FRED}
\usage{
FRED2(series, key)
}
\arguments{
\item{series}{text name of series}

\item{key}{Your FRED API key (can be obtained for free - check http://api.stlouisfed.org/api_key.html)}
}
\value{
data frame
}
\description{
\code{FRED2} returns a data frame with series data
}
\details{
Queries the FRED API to get the requested data
}
\keyword{FRED2}

