% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FRED.r
\name{FREDcat}
\alias{FREDcat}
\title{Get info on variables in one or more FRED categories}
\usage{
FREDcat(catvec, key)
}
\arguments{
\item{catvec}{numeric vector with list of FRED categories to get info for}

\item{key}{Your FRED API key (can be obtained for free - check http://api.stlouisfed.org/api_key.html)
default returns data frame with info on series available in one or several categories}
}
\value{
data frame
}
\description{
\code{FREDcat} returns a data frame with info on series
}
\details{
Queries the FRED API to get the requested data
}
\examples{
\dontrun{
require(dplyr)
df <- FREDcat(c(106, 125))
df <- FREDcat(18) # 18 is NIPA, so this will get info on ALL series in the NIPA category and its children
df2 <- df \%>\% filter(grepl("Billions of Chained 2009 Dollars", units),
              grepl("consumption", title, ignore.case=TRUE),
              grepl("durable", title, ignore.case=TRUE)) \%>\%
       select(cat, id, title, frequency_short, seasonal_adjustment_short, units_short)
# Look at df2 in the viewer to find the desired consumption variables
# Let's look at PCDGCC96 real quarterly durables, and PCNDGC96 real q nondurables
head(df)
count(df, id, title)
vars <- c("PCDGCC96", "PCNDGC96")
FRED(vars)
}
}
\keyword{FREDcat}

