% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FRED_qm.r
\name{FREDqm}
\alias{FREDqm}
\title{Query FRED via quantmod to get a single variable}
\usage{
FREDqm(var, freq)
}
\arguments{
\item{var}{a variable name in c("gdp", "gdppi", "rgdp")}

\item{freq}{"q" or "a"; default: "q"}
}
\value{
data frame with data
}
\description{
\code{FREDqm} returns a data frame with requested regional data. Right now only works with annual.
}
\details{
Queries FRED via quantmod to get the data.
}
\examples{
\dontrun{
FREDqm()
df <- FREDqm("gdppi", "q")
tail(df)
tail(NIPA("gdppi", "q"))
df <- FREDqm("gdp", "a")
tail(df)
tail(NIPA("gdp", "a"))
df <- FREDqm("rgdp")
tail(df)
tail(NIPA("rgdp"))
}
}
\keyword{FREDqm}

