% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FOF.r
\name{getz1df.fromfile}
\alias{getz1df.fromfile}
\title{Get ALL Federal Reserve Board Z1 Flow of Funds data from previously downloaded zip file}
\usage{
getz1df.fromfile(z1dir)
}
\arguments{
\item{z1dir}{defines a directory where the Z1 zip file will be stored; default is tempdir()}
}
\value{
data frame
}
\description{
\code{getz1df.fromfile} Downloads latest Z1 data from the FRB site (http://www.federalreserve.gov/datadownload)
and returns a data frame with all Z1 data
}
\details{
Downloads the data, reads it with scan, parses data, and cleans data
}
\examples{
\dontrun{
z1 <- getz1df.fromfile()
head(z1)
# or from existing fof dir: z1 <- getz1df.fromfile(fof) where fof <- "e.Data.FOF."
}
}
\keyword{getz1df.fromfile}

