% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{NIPA_DataMult}
\alias{NIPA_DataMult}
\title{Get a Data Frame With Multiple BEA Tables}
\usage{
NIPA_DataMult(nipatablenames, freq, dsname, key)
}
\arguments{
\item{nipatablenames}{a character vector with NIPA table numbers (see example below); default: no default}

\item{freq}{"q" or "a"; default: "q"}

\item{dsname}{text name of the dataset (e.g., "NIPA") (see \code{BEA_DSlist}); default: "NIPA"}

\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov); default: my key}
}
\value{
vector with TableIDs
}
\description{
\code{NIPA_DataMult} returns a data frame with multiple BEA tables
}
\details{
Queries the BEA API to get the data
}
\examples{
require(dplyr)
nipatabvec <- c("2.1", "2.3.4", "2.3.6", "3.3", "3.9.6", "3.10.6")
df <- NIPA_DataMult(nipatabvec) # this can take a while
head(df)
count(df, TableID, Description)
}
\keyword{NIPA_DataMult}

