% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BEA_api.r
\name{NIPATableIDs}
\alias{NIPATableIDs}
\title{Get a Vector of TableIDs Corresponding to NIPA Table Numbers}
\usage{
NIPATableIDs(nipatabvec, dsname, key)
}
\arguments{
\item{nipatabvec}{character vector with partial NIPA table names}

\item{dsname}{text name of the dataset (e.g., "NIPA") (see \code{BEA_DSlist})}

\item{key}{Your BEA API key (can be obtained for free - check www.bea.gov)}
}
\value{
vector with TableIDs
}
\description{
\code{NIPATableIDs} returns a vector of TableIDs corresponding to NIPA table numbers
}
\details{
Queries the BEA API to get a list of tables for a particular dsname, and then finds the TableIDs for the given NIPA tables
}
\examples{
nipatabvec <- c("2.1", "2.3.4", "2.3.6", "3.3", "3.9.6", "3.10.6")
NIPATableIDs(nipatabvec)
}
\keyword{NIPATableIDs}

