% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/information.R
\name{Param}
\alias{Param}
\alias{info}
\alias{intro}
\alias{link}
\alias{origin}
\title{Get the information of the APIfunction}
\usage{
Param(object)
intro(object)
info(object)
link(object)
origin(object)
}
\arguments{
\item{object}{an APIfunction object created by e.g. \code{APIstore}}
}
\value{
the information of the Object
}
\description{
Get the information of the APIfunction which created by
\code{APIstore}
}
\details{
\code{Param} show the Parameter information of the function.

\code{intro} show the Introduction information of the function.
\code{info} show the information of function return .
\code{link} show the link of the function.
\code{origin} show the origin of the function.
}
\examples{
function440 = APIStore(440)
Param(function440)
intro(function440)
info(function440)
link(function440)
origin(function440)
}
\author{
Chiffon <\url{http://chiffon.gitcafe.io}>
}

