% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FOF.r
\name{z1all}
\alias{z1all}
\title{Get latest version of ALL Federal Reserve Board Z1 Flow of Funds data from web}
\usage{
z1all(z1dir)
}
\arguments{
\item{z1dir}{defines a directory where the Z1 zip file will be stored; default is tempdir()}
}
\value{
data frame
}
\description{
\code{z1all} Downloads latest Z1 data from the FRB site (http://www.federalreserve.gov/datadownload)
and returns a data frame with all Z1 data
}
\details{
Downloads the data, reads it with scan, parses data, and cleans data
}
\examples{
z1 <- z1all()
head(z1)
}
\keyword{z1all}

