% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FRED_qm.r
\name{FRED}
\alias{FRED}
\title{Query FRED via quantmod to get a single variable}
\usage{
FRED(var, freq)
}
\arguments{
\item{var}{a variable name in c("gdp", "gdppi", "rgdp")}

\item{freq}{"q" or "a"; default: "q"}
}
\value{
data frame with data
}
\description{
\code{FRED} returns a data frame with requested regional data. Right now only works with annual.
}
\details{
Queries FRED via quantmod to get the data.
}
\examples{
FRED()
df <- FRED("gdppi", "q")
tail(df)
tail(NIPA("gdppi", "q"))
df <- FRED("gdp", "a")
tail(df)
tail(NIPA("gdp", "a"))
df <- FRED("rgdp")
tail(df)
tail(NIPA("rgdp"))
}
\keyword{FRED}

