% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FRED.r
\name{FRED}
\alias{FRED}
\title{Get one or more series from FRED}
\usage{
FRED(svec, key, full)
}
\arguments{
\item{svec}{character vector with list of series}

\item{key}{Your FRED API key (can be obtained for free - check http://api.stlouisfed.org/api_key.html)}

\item{full}{determines whether to return all data including realtime dates
default returns series name, date, and value}
}
\value{
data frame
}
\description{
\code{FRED} returns a data frame with series data
}
\details{
Queries the FRED API to get the requested data

"Nicknames" are allowed for certain commonly used series:

   gdp.q, gdp.a         for nominal gdp, quarterly or annual

   gdppi.q, gdppi.a     for gdp price index, quarterly or annual

   rgdp.q, rgdp.a       for real gdp, quarterly or annual
}
\examples{
require(dplyr)
df <- FRED("GDP")
head(df)
df <- FRED(c("GDP", "GNP"))
tail(df)
vlist <- c("GDP", "GNP", "GDPCTPI", "GNPCA")
df <- FRED(vlist)
tail(df)
count(df, series)
count(df, freq)
}
\keyword{FRED}

